+++ src/environments/environment.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: false,
-  version: '1.1.1',
+  version: '1.1.2',
   curseForgeApiKey: ''
 };